requirements.txt
streamlit
matplotlib
pillow
tensorflow 
numpy
pandas 
importlib
gdown